+++ dashboard/static/assets/js/components/core/App.js
@@ -1,54 +1,54 @@
-import React from 'react';
+import React from "react";
 
-import { Navbar } from './coreComponents/navbar';
+import Navbar from "./coreComponents/navbar";
 
-import { Home } from './coreComponents/home';
-import { Resources } from './coreComponents/resources';
-import { Mutual } from './coreComponents/mutual';
-import { Contact } from './coreComponents/contact';
+import Home from "./coreComponents/home"; // FIXME: proper path
+import { Resources } from "./coreComponents/resources";
+import { Mutual } from "./coreComponents/mutual";
+import { Contact } from "./coreComponents/contact";
 
-import { ClientAuthenticationView } from '../common/accounts/base'
+import { ClientAuthenticationView } from "../common/accounts/base";
 
 export function HomePageView() {
-    return (
-        <>
-            <Navbar />
-            <Home />
-        </>
-    );
+  return (
+    <>
+      <Navbar />
+      <Home />
+    </>
+  );
 }
 
 export function ResourcesPageView() {
-    return (
-        <>
-            <Navbar />
-            <Resources />
-        </>
-    );
+  return (
+    <>
+      <Navbar />
+      <Resources />
+    </>
+  );
 }
 
 export function MutualPageView() {
-    return (
-        <>
-            <Navbar />
-            <Mutual />
-        </>
-    );
+  return (
+    <>
+      <Navbar />
+      <Mutual />
+    </>
+  );
 }
 
 export function ContactPageView() {
-    return (
-        <>
-            <Navbar />
-            <Contact />
-        </>
-    );
+  return (
+    <>
+      <Navbar />
+      <Contact />
+    </>
+  );
 }
 
 export function ClientAuthView({ formType }) {
-    return (
-        <>
-            <ClientAuthenticationView formType={formType} />
-        </>
-    )
+  return (
+    <>
+      <ClientAuthenticationView formType={formType} />
+    </>
+  );
 }
