+++ dashboard/static/assets/js/components/core/coreComponents/home.js
@@ -1,23 +1,11 @@
-import React from 'react';
+import React from 'react'
 
-
-import { Hero } from './home/hero';
-import { Disclaimer } from './home/disclaimer';
-import { WhyUsSectionWrapper } from './home/whyus';
-import { PersonHero } from './home/personHero';
-import { Footer } from './footer';
-
-export function Home() {
-    
-    document.body.style.overflowY = "auto";
-
-    return (
-        <>
-            <Hero />
-            <WhyUsSectionWrapper />
-            <PersonHero />
-            <Disclaimer />
-            <Footer />
-        </>
-    );
+const home = () => {
+  return (
+    <div>
+      <h1 className='text-3xl text-center text-red-700'>checking tailwind home</h1>
+    </div>
+  )
 }
+
+export default home
