+++ dashboard/templates/includes/head.html
@@ -50,7 +50,7 @@
 
     body{
         font-family: 'Inter', sans-serif !important;
-        background-color: var(--main-color) !important;
+        {% comment %} background-color: var(--main-color) !important; {% endcomment %}
         font-size: 0.9rem !important;
     }
